Actor icons load from sprites

+++ OpenRA.ModMaker/Model/SpriteImage.cs
@@ -4,7 +4,7 @@ namespace OpenRA.ModMaker.Model
 {
 	public class SpriteImage
 	{
-		public SpriteImage(int width, int height, Lazy<byte[]> sheetData)
+		public SpriteImage(int width, int height, Lazy<int[]> sheetData)
 		{
 			Width = width;
 			Height = height;
@@ -13,6 +13,6 @@ namespace OpenRA.ModMaker.Model
 
 		public int Width { get; set; }
 		public int Height { get; set; }
-		public Lazy<byte[]> RawData { get; private set; }
+		public Lazy<int[]> RawData { get; private set; }
 	}
 }
